requirements.txt
discord.py==2.5.2
requests==2.32.4
aiohttp==3.12.15
